clutch-backend-zip-contents/requirements.txt
fastapi
uvicorn
SpeechRecognition==3.10.0
watchdog==3.0.0
pydub==0.25.1
openai>=1.0.0
python-dotenv==1.0.0
nltk==3.8.1
matplotlib==3.7.1
wordcloud==1.9.2
requests>=2.26.0
streamlit>=1.24.0
flask>=2.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pytesseract>=0.3.0
boto3>=1.39.0
discord.py[voice]>=2.3.0
google-cloud-texttospeech>=2.14.0
mutagen>=1.47.0
